LeaseTime 초과 시 트랜잭션 롤백하도록 수정 (#46)

+++ Tradin-Core/src/main/java/com/tradin/core/common/annotation/DistributedLock.java
@@ -28,7 +28,7 @@ public @interface DistributedLock {
     /**
      * 락 임대 시간 (default - 3s) 락을 획득한 이후 leaseTime 이 지나면 락을 해제한다
      */
-    long leaseTime() default 3L;
+    int leaseTime() default 3;
 
     String fallbackMethod() default "";
